Default to JSON statistics channel
This also drops support for "auto" determination of stats version, which
is too tightly coupled to the XML stats channel to easily support JSON
too. It is essentially redundant anyway, since the JSON stats channel is
supported and always present since BIND v9.10.0.

We also no longer unmarshal the memory or socketmgr nodes of the XML
stats, since we don't currently expose any metrics for these, and they
have been responsible for unmarshal errors in the past. Once we actually
_do_ expose metrics for them, they should be included in tests.

Signed-off-by: Daniel Swarbrick <daniel.swarbrick@gmail.com>